refactor room page into modular components
